+++ chat-agent/dev/src/app/(payload)/admin/importMap.js
@@ -21,8 +21,8 @@ import { StrikethroughFeatureClient as StrikethroughFeatureClient_e70f5e05f09f93
 import { UnderlineFeatureClient as UnderlineFeatureClient_e70f5e05f09f93e00b997edb1ef0c864 } from '@payloadcms/richtext-lexical/client'
 import { BoldFeatureClient as BoldFeatureClient_e70f5e05f09f93e00b997edb1ef0c864 } from '@payloadcms/richtext-lexical/client'
 import { ItalicFeatureClient as ItalicFeatureClient_e70f5e05f09f93e00b997edb1ef0c864 } from '@payloadcms/richtext-lexical/client'
-import { ChatNavLinkServer as ChatNavLinkServer_58e4688362c6fde9fcb5d132423d2812 } from '@jhb.software/payload-chat-agent'
-import { ChatViewServer as ChatViewServer_58e4688362c6fde9fcb5d132423d2812 } from '@jhb.software/payload-chat-agent'
+import { ChatNavLinkServer as ChatNavLinkServer_48b989653b39a67494f0d80b0f9fbeb0 } from '@jhb.software/payload-chat-agent/server'
+import { ChatViewServer as ChatViewServer_48b989653b39a67494f0d80b0f9fbeb0 } from '@jhb.software/payload-chat-agent/server'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 export const importMap = {
@@ -49,7 +49,7 @@ export const importMap = {
   "@payloadcms/richtext-lexical/client#UnderlineFeatureClient": UnderlineFeatureClient_e70f5e05f09f93e00b997edb1ef0c864,
   "@payloadcms/richtext-lexical/client#BoldFeatureClient": BoldFeatureClient_e70f5e05f09f93e00b997edb1ef0c864,
   "@payloadcms/richtext-lexical/client#ItalicFeatureClient": ItalicFeatureClient_e70f5e05f09f93e00b997edb1ef0c864,
-  "@jhb.software/payload-chat-agent#ChatNavLinkServer": ChatNavLinkServer_58e4688362c6fde9fcb5d132423d2812,
-  "@jhb.software/payload-chat-agent#ChatViewServer": ChatViewServer_58e4688362c6fde9fcb5d132423d2812,
+  "@jhb.software/payload-chat-agent/server#ChatNavLinkServer": ChatNavLinkServer_48b989653b39a67494f0d80b0f9fbeb0,
+  "@jhb.software/payload-chat-agent/server#ChatViewServer": ChatViewServer_48b989653b39a67494f0d80b0f9fbeb0,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }